Add files via upload
This model has been converted to .pb form, and is from a technical viewpoint, converted to an openVINO format. However it has been notified that the openVINO optimization is not visible in all tested devices. This does not inhibit the access to the file or affect the general execution of TensorFlow model